// frontend/src/interfaces/ifVehicles.ts
export interface VehiclestInterface {
  ID: any;
  BusID?: number;
  Type?: string;           
  BusRegistration?: string;
  Brand?: string;
  Series?: string;
  Year?: number;         
  EngineAndPower?: string; 
  FuelTank?: number;       
  Seat?: number;       

}
/*เอาไว้ match ชนิดข้อมูลเฉยๆ*/