// frontend/src/interfaces/ifEmployees.ts
export interface EmployeesInterface {
  ID: any;
  EmployeeID?: number;
  Username?:  string;
  Password?:  string;
  FirstName?: string;
  LastName?: string;
  PhoneNumber?:string;
  Position?: string;
  Address?: string;
  GenderID?: number;
}
/*เอาไว้ match ชนิดข้อมูลเฉยๆ*/